+++ packages/gafl-webapp-service/src/handlers/__tests__/cancel-recurring-payment-authentication-handler.spec.js
@@ -1,5 +1,11 @@
 import handler from '../cancel-recurring-payment-authentication-handler'
-import { CANCEL_RP_IDENTIFY, CANCEL_RP_DETAILS, CANCEL_RP_AGREEMENT_NOT_FOUND } from '../../uri.js'
+import {
+  CANCEL_RP_IDENTIFY,
+  CANCEL_RP_DETAILS,
+  CANCEL_RP_AGREEMENT_NOT_FOUND,
+  CANCEL_RP_ALREADY_CANCELLED,
+  CANCEL_RP_LICENCE_NOT_FOUND
+} from '../../uri.js'
 import { salesApi } from '@defra-fish/connectors-lib'
 
 jest.mock('../../processors/uri-helper.js')
@@ -19,7 +25,9 @@ jest.mock('@defra-fish/business-rules-lib', () => ({
 jest.mock('../../uri.js', () => ({
   CANCEL_RP_IDENTIFY: { page: 'cancel-rp-identify page', uri: Symbol('cancel-rp-identify-uri') },
   CANCEL_RP_DETAILS: { uri: Symbol('cancel-rp-details-uri') },
-  CANCEL_RP_AGREEMENT_NOT_FOUND: { uri: Symbol('cancel-rp-agreement-not-found-uri') }
+  CANCEL_RP_AGREEMENT_NOT_FOUND: { uri: Symbol('cancel-rp-agreement-not-found-uri') },
+  CANCEL_RP_ALREADY_CANCELLED: { uri: Symbol('cancel-rp-already-cancelled-uri ') },
+  CANCEL_RP_LICENCE_NOT_FOUND: { uri: Symbol('cancel-rp-licence-not-found-uri') }
 }))
 jest.mock('../../processors/recurring-payments-write-cache.js')
 
@@ -99,9 +107,9 @@ describe('Cancel RP Authentication Handler', () => {
   })
 
   describe('Unsuccessful authentication - no match', () => {
-    it('redirects to the CANCEL_RP_IDENTIFY.uri', async () => {
+    it('redirects to the CANCEL_RP_LICENCE_NOT_FOUND.uri', async () => {
       const { h } = await invokeHandlerWithMocks({ salesApiResponse: null })
-      expect(h.redirectWithLanguageCode).toHaveBeenCalledWith(CANCEL_RP_IDENTIFY.uri)
+      expect(h.redirectWithLanguageCode).toHaveBeenCalledWith(CANCEL_RP_LICENCE_NOT_FOUND.uri)
     })
 
     it('returns value of redirect', async () => {
@@ -121,7 +129,7 @@ describe('Cancel RP Authentication Handler', () => {
         CANCEL_RP_IDENTIFY.page,
         expect.objectContaining({
           payload: expect.any(Object),
-          error: { referenceNumber: 'not-found' }
+          errorRedirect: true
         })
       )
     })
@@ -165,12 +173,12 @@ describe('Cancel RP Authentication Handler', () => {
     })
   })
 
-  describe('Unsuccessful authentication - RCP cancelled', () => {
-    it('redirects to the CANCEL_RP_IDENTIFY.uri', async () => {
+  describe('Unsuccessful authentication - RCP already cancelled', () => {
+    it('redirects to the CANCEL_RP_ALREADY_CANCELLED.uri', async () => {
       const { h } = await invokeHandlerWithMocks({
         salesApiResponse: { permission: { id: 'perm-id' }, recurringPayment: { id: 'rcp-id', status: 1, cancelledDate: '2024-01-01' } }
       })
-      expect(h.redirectWithLanguageCode).toHaveBeenCalledWith(CANCEL_RP_IDENTIFY.uri)
+      expect(h.redirectWithLanguageCode).toHaveBeenCalledWith(CANCEL_RP_ALREADY_CANCELLED.uri)
     })
 
     it('returns value of redirect', async () => {
@@ -184,7 +192,7 @@ describe('Cancel RP Authentication Handler', () => {
       expect(result).toBe(redirectResult)
     })
 
-    it('sets page cache error for RCP cancelled', async () => {
+    it('sets page cache for RCP already cancelled', async () => {
       const { request } = await invokeHandlerWithMocks({
         salesApiResponse: { permission: { id: 'perm-id' }, recurringPayment: { id: 'rcp-id', status: 1, cancelledDate: '2024-01-01' } }
       })
@@ -192,7 +200,7 @@ describe('Cancel RP Authentication Handler', () => {
         CANCEL_RP_IDENTIFY.page,
         expect.objectContaining({
           payload: expect.any(Object),
-          error: { recurringPayment: 'rcp-cancelled' }
+          errorRedirect: true
         })
       )
     })