+++ packages/gafl-webapp-service/src/handlers/cancel-recurring-payment-authentication-handler.js
@@ -1,7 +1,15 @@
-import { CANCEL_RP_IDENTIFY, CANCEL_RP_DETAILS, CANCEL_RP_AGREEMENT_NOT_FOUND } from '../../src/uri.js'
+import {
+  CANCEL_RP_IDENTIFY,
+  CANCEL_RP_DETAILS,
+  CANCEL_RP_AGREEMENT_NOT_FOUND,
+  CANCEL_RP_ALREADY_CANCELLED,
+  CANCEL_RP_LICENCE_NOT_FOUND
+} from '../../src/uri.js'
 import { salesApi } from '@defra-fish/connectors-lib'
 import { validation } from '@defra-fish/business-rules-lib'
 import { setupCancelRecurringPaymentCacheFromAuthResult } from '../processors/recurring-payments-write-cache.js'
+import { cacheDateFormat } from '../../src/processors/date-and-time-display.js'
+import moment from 'moment-timezone'
 import Joi from 'joi'
 
 const applyAuthFailure = async (request, h, { pageData, redirectUri, statusData }) => {
@@ -30,12 +38,18 @@ const cancelRecurringPaymentAuthenticationHandler = async (request, h) => {
   }
 
   if (!authenticationResult) {
-    context.pageData.error = { referenceNumber: 'not-found' }
+    context.pageData.errorRedirect = true
+    context.redirectUri = CANCEL_RP_LICENCE_NOT_FOUND.uri
   } else if (!authenticationResult.recurringPayment) {
     context.pageData.errorRedirect = true
     context.redirectUri = CANCEL_RP_AGREEMENT_NOT_FOUND.uri
   } else if (authenticationResult.recurringPayment.cancelledDate) {
-    context.pageData.error = { recurringPayment: 'rcp-cancelled' }
+    context.pageData.errorRedirect = true
+    context.redirectUri = CANCEL_RP_ALREADY_CANCELLED.uri
+    context.pageData.payload = {
+      ...context.pageData.payload,
+      endDate: moment(authenticationResult.recurringPayment.endDate).format(cacheDateFormat)
+    }
   }
 
   if (context.pageData.error || context.pageData.errorRedirect) {