+++ packages/gafl-webapp-service/src/routes/journey-definition.js
@@ -35,7 +35,9 @@ import {
   CANCEL_RP_CONFIRM,
   CANCEL_RP_COMPLETE,
   CANCEL_RP_AGREEMENT_NOT_FOUND,
-  JOURNEY_GOAL
+  JOURNEY_GOAL,
+  CANCEL_RP_ALREADY_CANCELLED,
+  CANCEL_RP_LICENCE_NOT_FOUND
 } from '../uri.js'
 
 import { CommonResults, CONTACT_SUMMARY_SEEN, ShowDigitalLicencePages } from '../constants.js'
@@ -507,6 +509,18 @@ export default [
   },
   {
     current: CANCEL_RP_AGREEMENT_NOT_FOUND,
+    backLink: CANCEL_RP_IDENTIFY.uri
+  },
+  {
+    current: CANCEL_RP_LICENCE_NOT_FOUND,
+    backLink: CANCEL_RP_IDENTIFY.uri
+  },
+  {
+    current: CANCEL_RP_ALREADY_CANCELLED,
+    backLink: CANCEL_RP_IDENTIFY.uri
+  },
+  {
+    current: CANCEL_RP_LICENCE_NOT_FOUND,
     backLink: CANCEL_RP_IDENTIFY.uri
   }
 ]