+++ packages/gafl-webapp-service/src/routes/telesales-routes.js
@@ -6,6 +6,8 @@ import cancelRPDetails from '../pages/recurring-payments/cancel/details/route.js
 import cancelRPConfirm from '../pages/recurring-payments/cancel/confirm/route.js'
 import cancelRPComplete from '../pages/recurring-payments/cancel/complete/route.js'
 import cancelRPAgreementNotFound from '../pages/recurring-payments/cancel/agreement-not-found/route.js'
+import cancelRPLicenceNotFound from '../pages/recurring-payments/cancel/licence-not-found/route.js'
+import cancelRPAlreadyCancelled from '../pages/recurring-payments/cancel/already-cancelled/route.js'
 
 const telesalesRoutes = [
   {
@@ -35,6 +37,14 @@ const telesalesRoutes = [
 ]
 
 if (process.env.SHOW_CANCELLATION_JOURNEY === 'true') {
-  telesalesRoutes.push(...cancelRPIdentify, ...cancelRPDetails, ...cancelRPConfirm, ...cancelRPComplete, ...cancelRPAgreementNotFound)
+  telesalesRoutes.push(
+    ...cancelRPIdentify,
+    ...cancelRPDetails,
+    ...cancelRPConfirm,
+    ...cancelRPComplete,
+    ...cancelRPAgreementNotFound,
+    ...cancelRPAlreadyCancelled,
+    ...cancelRPLicenceNotFound
+  )
 }
 export default telesalesRoutes