+++ packages/gafl-webapp-service/src/uri.js
@@ -62,6 +62,9 @@ export const ERROR_TESTING = { uri: '/buy/throw-error' }
 export const CHOOSE_PAYMENT = { uri: '/buy/choose-payment', page: 'choose-payment' }
 export const SET_UP_PAYMENT = { uri: '/buy/set-up-recurring-card-payment', page: 'set-up-payment' }
 
+/**
+ * Recurring payments cancellation pages
+ */
 export const CANCEL_RP_IDENTIFY = { uri: '/buy/cancel-recurring-payment/identify', page: 'cancel-rp-identify' }
 export const CANCEL_RP_AUTHENTICATE = { uri: '/buy/cancel-recurring-payment/authenticate' }
 export const CANCEL_RP_DETAILS = { uri: '/buy/cancel-recurring-payment/details', page: 'cancel-rp-details' }
@@ -71,6 +74,8 @@ export const CANCEL_RP_AGREEMENT_NOT_FOUND = {
   uri: '/buy/cancel-recurring-payment/agreement-not-found',
   page: 'cancel-rp-agreement-not-found'
 }
+export const CANCEL_RP_LICENCE_NOT_FOUND = { uri: '/buy/cancel-recurring-payment/licence-not-found', page: 'licence-not-found' }
+export const CANCEL_RP_ALREADY_CANCELLED = { uri: '/buy/cancel-recurring-payment/already-cancelled', page: 'already-cancelled' }
 
 export const JOURNEY_GOAL = { uri: '/buy/journey-goal', page: 'journey-goal' }